refactor: post 파일 삭제, 상수 분리
