+++ src/client/pages/randomEventResultPage/RandomEventResultPage.tsx
@@ -1,27 +1,27 @@
 import React, { useEffect, useRef, useState } from "react";
-import { useLocation, useNavigate } from "react-router";
+import { useNavigate } from "react-router";
+import { useErrorBoundary } from "react-error-boundary";
+import Cookies from "js-cookie";
 import { useAppContext } from "../../providers/AppProvider";
-import { postEvent2Answers } from "../../api/post";
+import { checkAndRefreshToken, postEvent2Answers } from "../../api/fetch";
+import { getDescriptionList } from "../../utils/util";
+import GlowBackground from "../../components/common/GlowBackground/GlowBackground";
+import HomeButton from "../../components/common/HomeButton/HomeButton";
+import AuthTooltip from "../../components/common/AuthTooltip/AuthTooltip";
+import Roulette from "../../components/randomEventPage/Roulette/Roulette";
 import RandomMainSection from "./RandomMainSection/RandomMainSection";
 import RandomExpectations from "./RandomExpectations/RandomExpectations";
-import Roulette from "../../components/randomEventPage/Roulette/Roulette";
 import car from "../../../common/assets/images/mainCar.webp";
-import { DRIVER_TYPE_LIST } from "../../constants/RandomEventData";
 import {
   AnswerInterface,
   RandomQuizResponseInterface,
 } from "../../types/RandomEvent";
-import { getDescriptionList } from "../../utils/util";
-import { useErrorBoundary } from "react-error-boundary";
-import { checkAndRefreshToken } from "../../api/fetch";
-import Cookies from "js-cookie";
-import GlowBackground from "../../components/common/GlowBackground/GlowBackground";
-import HomeButton from "../../components/common/HomeButton/HomeButton";
-import AuthTooltip from "../../components/common/AuthTooltip/AuthTooltip";
-
-const ANIMATION_DURATION = 5000;
-const ROULETTE_END_DELAY = 6000;
-const HEADER_TEXT = "당신의 운전자 유형은?";
+import {
+  ANIMATION_DURATION,
+  DRIVER_TYPE_LIST,
+  HEADER_TEXT,
+  ROULETTE_END_DELAY,
+} from "../../constants/RandomEventData";
 
 const ROULETTE_END_CONTAINER_CLASSES = {
   true: "justify-start pt-[16rem]",
@@ -34,23 +34,31 @@ const ROULETTE_END_HEADER_CLASSES = {
 };
 
 const RandomEventResultPage = () => {
-  const appContext = useAppContext();
+  const { isAuth, setIsAuth, isRandomEnd } = useAppContext();
   const navigate = useNavigate();
-  const location = useLocation();
-  const [timeElapsed, setTimeElapsed] = useState(0);
   const { showBoundary } = useErrorBoundary();
+  const [timeElapsed, setTimeElapsed] = useState(0);
+  const [answer, setAnswer] = useState<AnswerInterface | null>(null);
   const [resultData, setResultData] =
     useState<RandomQuizResponseInterface | null>(null);
   const randomExpectationsRef = useRef<HTMLDivElement>(null);
-
-  const { isAuth, setIsAuth, isRandomEnd } = appContext;
-  const [answer, setAnswer] = useState<AnswerInterface | null>(null);
 
   const isRouletteEnd = timeElapsed >= ROULETTE_END_DELAY;
   const animation = timeElapsed >= ANIMATION_DURATION;
 
   const containerStyle = ROULETTE_END_CONTAINER_CLASSES[`${isRouletteEnd}`];
   const headerStyle = ROULETTE_END_HEADER_CLASSES[`${isRouletteEnd}`];
+
+  const checkToken = async () => {
+    const response = await checkAndRefreshToken();
+    if (response.code === 200) {
+      setIsAuth(true);
+      Cookies.set("auth", response.result.accessToken, { expires: 1 / 24 });
+    } else {
+      setIsAuth(false);
+      Cookies.remove("auth");
+    }
+  };
 
   useEffect(() => {
     const interval = setInterval(() => {
@@ -97,20 +105,7 @@ const RandomEventResultPage = () => {
     };
 
     tryFetch();
-  });
-
-  const checkToken = async () => {
-    const response = await checkAndRefreshToken();
-    if (response.code === 200) {
-      setIsAuth(true);
-      Cookies.set("auth", response.result.accessToken, { expires: 1 / 24 });
-    } else {
-      setIsAuth(false);
-      Cookies.remove("auth");
-    }
-  };
-
-  if (resultData === null) return null;
+  }, []);
 
   return (
     <div className="flex flex-col items-center transition-all duration-500">
@@ -119,7 +114,7 @@ const RandomEventResultPage = () => {
       <GlowBackground />
       {resultData && (
         <div
-          className={`relative flex h-screen w-screen flex-col items-center ${containerStyle} gap-16 overflow-scroll transition-all duration-300`}
+          className={`relative flex h-screen w-screen flex-col items-center ${containerStyle} gap-16 transition-all duration-300`}
         >
           <div
             className={`absolute flex w-[36.5rem] flex-col items-center duration-500 ${headerStyle}`}
